Avoid loading page chooser into main page
